Notify user when restricted from running in background
Add a dialog box to app open when two conditions are met: the user has explicitly enabled notifications and the app is restricted by the system from running in the background. The user will only see this once per opening of the app, and then only if they take no action.