Add option to show less notifications
This option replaces the most recent date and time the background task has checked for updates with the date and time the user navigated to or from the main page. This means that, when the background task does check for updates, it will post less notifications if the app has been used in the intervening period. This assumes that the user looked at all articles they want to see from the time period they were using the app.

Also fixed a possible bug where the user could turn off notifications, freezing the last check date and time, then turn them on at some point in the future, resulting in a lot of notifications for that user.